fix: queries; chore

+++ src/modules/PaymentForm/queries/index.ts
@@ -1,7 +1,7 @@
 import { gql } from "@apollo/client";
 
 export const CREATE_TRANSACTION = gql`
-mutation createTransaction($order_id: ID!, $full_price: Int!, $order_items: [UpdatedOrderItem]!, $city: String!, $street: String!, $building: String!) {
+mutation createTransaction($order_id: ID!, $full_price: Float!, $order_items: [UpdatedOrderItem]!, $city: String!, $street: String!, $building: String!) {
     createTransaction(order_id: $order_id, full_price: $full_price, order_items: $order_items, city: $city, street: $street, building: $building) {
             id, full_price
         }
