+++ src/modules/Products/api/queries.ts
@@ -2,7 +2,7 @@ import { gql } from '@apollo/client'
 
 
 export  const GET_SORTED_PRODUCTS = gql` 
-    query getSortedProducts ($price: Int, $in_stock: Boolean, $discount: Boolean, $priceSort: String, $ratingSort: String, $showCount: Int, $category: ID) {
+    query getSortedProducts ($price: Float, $in_stock: Boolean, $discount: Boolean, $priceSort: String, $ratingSort: String, $showCount: Int, $category: ID) {
         sortedProducts (
                 price: $price, 
                 in_stock: $in_stock,
